Adjusted World Map to Stay Still
Made a script so that the world map will appear in front of the player when they open it and stay still after it's summoned in front of the player. Will go to player again when closed and reopened elsewhere.
Also added render textures to gitignore because they technically always get changed when the cameras update them